+++ Source/SwiftLintFramework/Rules/FirebaseCoreRule.swift
@@ -1,0 +1,57 @@
+//
+//  FirebaseCoreRule.swift
+//  SwiftLint
+//
+//  Created by Ibrahim Ulukaya on 3/8/17.
+//  Copyright © 2017 Realm. All rights reserved.
+//
+
+import SourceKittenFramework
+
+public struct FirebaseCoreRule: ConfigurationProviderRule, OptInRule {
+
+    public var configuration = SeverityConfiguration(.warning)
+
+    public init() {}
+
+    public static let description = RuleDescription(
+        identifier: "firebase_core",
+        name: "Firebase Core",
+        description: "Firebase should be configured before use.",
+        nonTriggeringExamples: [
+            "class AppDelegate: UIResponder, UIApplicationDelegate {\n" +
+            "  func application(_ application: UIApplication," +
+            "      didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {\n" +
+            "    FIRApp.configure()\n" +
+            "    return true \n" +
+            "  }\n" +
+            "}"
+        ],
+        triggeringExamples: [
+            "class AppDelegate: UIResponder, UIApplicationDelegate {\n" +
+            "  func application(_ application: UIApplication," +
+            "      didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {\n" +
+            "    return true \n" +
+            "  }\n" +
+            "}"
+        ]
+    )
+
+    public func validate(file: File) -> [StyleViolation] {
+        let appDelegate = file.structure.dictionary.substructure
+        if let first = appDelegate.first, first.inheritedTypes.contains("UIApplicationDelegate") {
+            for method in first.substructure where
+                SwiftDeclarationKind.functionMethodInstance.rawValue == method.kind &&
+                    method.name == "application(_:didFinishLaunchingWithOptions:)" {
+                for call in method.substructure where
+                    call.kind == SwiftExpressionKind.call.rawValue && call.name == "FIRApp.configure" {
+                    return []
+                }
+                return [StyleViolation(ruleDescription: type(of: self).description,
+                                      severity: configuration.severity,
+                                      location: Location(file: file, byteOffset: method.offset ?? 0))]
+            }
+        }
+        return []
+    }
+}